+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to Semantic Versioning (https://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Interposer pogo pads use the Mill-Max 806 footprint.
+
 ## [0.4.35] - 2026-08-24
 
 ### Added